pyproject.toml
[project]
name = "folder-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=4.0.0",
    "langchain-mcp>=0.2.1",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "mcp>=1.14.0",
    "pydantic>=2.11.9",
    "streamlit>=1.49.1",
    "python-dotenv>=1.0.0",
    "fastmcp>=0.1.0",
]
